% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaList.R
\name{aaList}
\alias{aaList}
\title{Return a vector with the 20 standard amino acids in upper case}
\usage{
aaList()
}
\value{
A character vector with the 20 standard amino acids in upper case.
}
\description{
This function returns a vector with the 20 standard amino acids in upper case.
}
\references{
Lu, Y., & Freeland, S. (2006). On the evolution of the standard amino-acid alphabet. Genome biology, 7(1), 102.
}
\author{
Richel Bilderbeek <richel@richelbilderbeek.nl>
}
