% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocovariance.R
\name{autoCovariance}
\alias{autoCovariance}
\title{Compute the auto-covariance index of a protein sequence}
\usage{
autoCovariance(sequence, lag, property, center = TRUE)
}
\description{
This function computes the Cruciani et al (2004) auto-corvariance index.
}
