% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasgaiVectors.R
\name{fasgaiVectors}
\alias{fasgaiVectors}
\title{Compute the FASGAI vectors of a protein sequence}
\usage{
fasgaiVectors(seq)
}
\arguments{
\item{seq}{An amino-acids sequence}
}
\value{

}
\description{
The FASGAI vectors (Factor Analysis Scales of Generalized Amino Acid Information) are based in involved hydrophobicity, alpha and turn propensities, bulky properties, composition characteristics, local flexibility, and electronic properties of 335 properties that can be utilized to represent the sequence structural features of peptides or protein motifs.
}
