\name{Peptides-package}
\alias{Peptides}
\docType{package}
\title{
Calculate indices and theoretical physicochemical properties of peptides and proteins sequences.
}
\description{
Calculate physicochemical properties and indices from aminoacid sequences of peptides 
and proteins. Include also utilities for read and plot GROMACS output files .XVG.
}
\details{
\tabular{ll}{
Package: \tab Peptides\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-02-06\cr
License: \tab GPL\cr
}
}
\author{
Daniel Osorio, Paola Rondon - Villarreal and Rodrigo Torres PhD.

Maintainer: Daniel Osorio <daniel.osorio@correo.uis.edu.co>
}