% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.xvg.R
\name{plot.xvg}
\alias{plot.xvg}
\title{Plot time series from GROMACS XVG files}
\usage{
\method{plot}{xvg}(x, ...)
}
\arguments{
\item{x}{A .XVG output file of the GROMACS molecular dynamics package}
}
\description{
Read and plot output data from a XVG format file.
}
\examples{
file <- system.file("xvg-files/epot.xvg",package="Peptides")
plot.xvg(file)
}
\author{
Daniel Osorio <daniel.osorio@correo.uis.edu.co>
}
\references{
Pronk, S., Pall, S., Schulz, R., Larsson, P., Bjelkmar, P., Apostolov, R., ... & Lindahl, E. (2013). GROMACS 4.5: a high-throughput and highly parallel open source molecular simulation toolkit. Bioinformatics, 29 (7), 845-854.
}

