% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-Z.R
\docType{data}
\name{Z}
\alias{Z}
\title{The 5 Z-scales from Sandberg et al. (1998)}
\format{A list with the five Sandberg et al. (1998) Z-scales. Each Z-scale represent an amino-acid property as follows: #' \itemize{
\item{Z1:} Lipophilicity
\item{Z2:} Steric properties (Steric bulk/Polarizability)
\item{Z3:} Electronic properties (Polarity / Charge) 
\item{Z4 and Z5:} They relate electronegativity, heat of formation, electrophilicity and hardness.
}}
\usage{
data(Z)
}
\description{
The five Sandberg et al. (1998) Z-scales describe each amino acid with numerical values, descriptors, which represent the physicochemical properties of the amino acids including NMR data and thin-layer chromatography (TLC) data.
}
\references{
Sandberg M, Eriksson L, Jonsson J, Sjostrom M, Wold S: New chemical descriptors relevant for the design of biologically active peptides. A multivariate characterization of 87 amino acids. J Med Chem 1998, 41:2481-2491.
}
