% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosscovariance.R
\name{crossCovariance}
\alias{crossCovariance}
\title{Compute the cross-covariance index of a protein sequence}
\usage{
crossCovariance(sequence, lag, property1, property2, center = TRUE)
}
\description{
This function computes the Cruciani et al (2004) cross-covariance index.
}
