\name{Peptides-package}
\alias{Peptides}
\docType{package}
\title{
Calculate indices and theoretical physicochemical properties of peptides and proteins sequences.
}
\description{
Calculate physicochemical properties and indices from amino acid sequences of peptides 
and proteins. Include also utilities for read and plot GROMACS output files .XVG.
}
\details{
\tabular{ll}{
Package: \tab Peptides\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-30\cr
License: \tab GPL-2\cr
}
}
\author{
Daniel Osorio, Paola Rondon-Villarreal and Rodrigo Torres.

Maintainer: Daniel Osorio <daniel.osorio@correo.uis.edu.co>
}
