\name{Peptides-package}
\alias{Peptides}
\docType{package}
\title{
Calculate Indices and Theoretical Properties of Protein Sequences
}
\description{
Calculate physicochemical properties and indices from amino-acid sequences of peptides and proteins. Include also the option to read and plot output files from the 'GROMACS' molecular dynamics package.
}
\details{
\tabular{ll}{
Package: \tab Peptides\cr
Type: \tab Package\cr
Version: \tab 1.1.1\cr
Date: \tab 2015-11-26\cr
License: \tab GPL-2\cr
}
}
\author{
Daniel Osorio, Paola Rondon-Villarreal and Rodrigo Torres.

Maintainer: Daniel Osorio <daniel.osorio@correo.uis.edu.co>
}
